Development

+++ RngHelpdesk.Contracts/Points/Queries/GetPointHistoryForUserQuery.cs
@@ -2,5 +2,5 @@
 
 public sealed class GetPointHistoryForUserQuery
 {
-    public int UserId { get; init; }
+    public ulong UserId { get; init; }
 }
