+++ RngHelpdesk.Contracts/Users/Commands/CreateUserResponse.cs
@@ -2,7 +2,7 @@
 
 public sealed class CreateUserResponse
 {
-    public int UserId { get; init; }
+    public ulong UserId { get; init; }
     public string Username { get; init; } = string.Empty;
     public string TemporaryPassword { get; init; } = string.Empty;
 }
